% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ens_weights_no_intercept}
\alias{ens_weights_no_intercept}
\title{Ensemble pathway weights}
\format{
a matrix containing weights
\describe{
}
}
\usage{
ens_weights_no_intercept
}
\description{
Weights for signatures from each cell line for each pathway.
Weights were determined by logistic regression (with LASSO) using LINCS
chemical perturbations of pathway nodes.
}
\keyword{datasets}
