% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_meta}
\alias{gene_meta}
\title{Gene metadata for all 978 measured L1000 genes}
\format{
A data frame containing Entrez gene IDs, HUGO gene symbols, full
gene names, and probe IDs for the L1000 genes
\describe{
}
}
\usage{
gene_meta
}
\description{
A data set containing metadata for all genes used in iPAS signatures
}
\keyword{datasets}
