% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iPAS_bar.R
\name{iPAS_bar}
\alias{iPAS_bar}
\title{Produce bar plots of top iPAS pathways}
\usage{
iPAS_bar(
  res,
  experiment = 1,
  num_top = 15,
  col_scale_max = 5,
  pal = "vikO",
  pathway_names = c("names", "identifiers", "both", "none"),
  short_names = TRUE,
  rev = FALSE,
  fill = "z_score",
  bar_height = "neg_log10_p",
  arrange_col = "p_emp",
  category = c("Signaling", "Cancer", "Disease", "Other", "All")
)
}
\arguments{
\item{res}{output object from the iPAS_enrich function}

\item{experiment}{the sample number or name of the sample from which to
graph results.}

\item{num_top}{the number of top pathways (by iPAS score) to plot}

\item{col_scale_max}{a positive number that defines the limits of the color
scale. The color scale is defined as c(-col_scale_max, col_scale_max).}

\item{pal}{scico package color palette. See here:
https://github.com/thomasp85/scico}

\item{pathway_names}{whether to show the names of KEGG pathways (e.g.
"mTOR signaling pathway - Homo sapiens (human)"), their unique identifiers
(e.g. "hsa04150") or both. "names" shows pathway names, "identifiers" shows
only ID numbers, "both" shows both, "none" shows no names.}

\item{short_names}{if TRUE, show shortened pathway names, not including
" - Homo sapiens (human)" at the end}

\item{rev}{T/F whether to reverse the order of the bars in the chart or not}

\item{fill}{the column to be used for the fill (color) of the bars. The
default is "z_score", but other options are "score" for the raw iPAS score
or "neg_log10_p" for the p-value (negative log10-transformed).}

\item{bar_height}{the column to be used for the height of the bars. The
choices are the same as above, the default is "neg_log10_p", the p-value
(negative log10-transformed).}

\item{arrange_col}{the column to use to order the bars, by default "p_emp"
for the empirical p-value (non-transformed).}

\item{category}{the categories of KEGG pathways for which to calculate iPAS
scores. We categorize pathways as "Disease", "Other", "Signaling", or
"Cancer". By default all categories/pathways are included.}
}
\description{
This function creates bar plots that graph iPAS results. The input to the
function is the object returned from the iPAS_enrich function.
}
