% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iPAS_density.R
\name{iPAS_density}
\alias{iPAS_density}
\title{Produce a density plot of iPAS permutation scores for a given pathway}
\usage{
iPAS_density(
  res,
  experiment,
  score_type = c("z_score", "score"),
  path,
  fill = "light blue"
)
}
\arguments{
\item{res}{output object from the iPAS_enrich function}

\item{experiment}{the sample number or name of the sample from which to
graph results.}

\item{score_type}{the type of iPAS score to graph. "score" will show raw iPAS
scores and "z_score" will show normalized iPAS scores.}

\item{path}{the KEGG pathway ID of the pathway to graph (e.g. "hsa04150" for
the mTOR pathway)}

\item{fill}{the color to fill in the density plot, "light blue" by default.}
}
\description{
This function returns a plot of the probability density of the null
distribution of iPAS scores for permutation signatures for a given pathway.
The iPAS score for the (non-permuted) input signature is shown by a red bar.
}
