% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iPAS_density_facet.R
\name{iPAS_density_facet}
\alias{iPAS_density_facet}
\title{Produce density plots of iPAS permutation scores for a given pathway for
multiple experiments}
\usage{
iPAS_density_facet(
  res,
  score_type = c("z_score", "score"),
  path,
  fill = "light blue"
)
}
\arguments{
\item{res}{output object from the iPAS_enrich function}

\item{score_type}{the type of iPAS score to graph. "score" will show raw iPAS
scores and "z_score" will show normalized iPAS scores.}

\item{path}{the KEGG pathway ID of the pathway to graph (e.g. "hsa04150" for
the mTOR pathway)}

\item{fill}{the color to fill in the density plot, "light blue" by default.}
}
\description{
This function is the same as the iPAS_density function, except for all
experiments at once.
The function returns a plot of the probability density of the null
distribution of iPAS scores for permutation signatures for a given pathway.
The iPAS score for the (non-permuted) input signature is shown by a red bar.
}
