% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iPAS_heatmap.R
\name{iPAS_heatmap}
\alias{iPAS_heatmap}
\title{Produce a heatmap of iPAS pathway scores}
\usage{
iPAS_heatmap(
  res,
  row_fontsize = 6,
  col_fontsize = 12,
  short_names = TRUE,
  pathway_names = c("names", "identifiers", "both"),
  category = c("Signaling", "Cancer", "Disease", "Other", "All")
)
}
\arguments{
\item{res}{output object from the iPAS_enrich function}

\item{row_fontsize}{font size for the row names (pathways), default is 6.}

\item{col_fontsize}{font size for the column names (input signatures),
default is 12.}

\item{short_names}{if TRUE, show shortened pathway names, not including
" - Homo sapiens (human)" at the end}

\item{pathway_names}{whether to show the names of KEGG pathways (e.g.
"mTOR signaling pathway - Homo sapiens (human)"), their unique identifiers
(e.g. "hsa04150") or both. "names" shows pathway names, "identifiers" shows
only ID numbers, "both" shows both, "none" shows no names.}

\item{category}{the categories of KEGG pathways for which to calculate iPAS
scores. We categorize pathways as "Disease", "Other", "Signaling", or
"Cancer". By default all categories/pathways are included.}
}
\description{
Produce a heatmap of iPAS pathway scores
}
