% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iPAS_list}
\alias{iPAS_list}
\title{Individual pathway activity signatures (iPAS)}
\format{
A list with 12 matrices one for each cell line
\describe{
  \item{columns}{204 KEGG pathways: hsa01521, hsa01522, etc.}
  \item{rows}{978 Entrez gene IDs (one for each L1000 probe): 10007, 1001, etc.}
}
}
\usage{
iPAS_list
}
\description{
A data set containing the integrative Pathway Activity Signatures (iPAS)
The variables are as follows:
}
\keyword{datasets}
