% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iPAS_score_perm.R
\name{iPAS_score_perm}
\alias{iPAS_score_perm}
\title{Create a null distribution of iPAS pathway scores using permutations}
\usage{
iPAS_score_perm(
  query,
  query_perm,
  pas,
  similarity,
  perm,
  pathway,
  cell_line,
  ens_weight,
  overlap_min
)
}
\description{
Create a null distribution of iPAS pathway scores using permutations
}
