% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kegg_meta}
\alias{kegg_meta}
\title{KEGG pathway metadata}
\format{
A data frame containing KEGG pathway metadata
\describe{
}
}
\usage{
kegg_meta
}
\description{
A data set containing metadata for each KEGG pathway.
}
\keyword{datasets}
