% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_iPAS.R
\name{load_iPAS}
\alias{load_iPAS}
\title{Load integrative Pathway Activity Signatures (iPAS)}
\usage{
load_iPAS(
  method = c("ensemble", "cl", "mtl", "legacy_cl"),
  cell_line = c("A375", "A549", "HA1E", "HCC515", "HEKTE", "HEPG2", "HT29", "MCF7",
    "NPC", "PC3", "SW480", "VCAP"),
  genes = c("entrez", "symbol")
)
}
\arguments{
\item{method}{The method used to create the iPAS signatures: "ensemble" refers to consensus signatures combined with weights determined by ensemble learning, "cl" refers to signatures created using data from one cell line only, "mtl" refers to "multi-task learning" signatures, "legacy_cl" refers to the version of signatures (from one cell line) published in Ren et al. 2020 (slightly different method than recent signatures).}

\item{cell_line}{LINCS cell line to load signatures from (if applicable)}

\item{genes}{format of rownames for genes (entrez ids or gene symbols)}
}
\value{
a list object containing the requested Pathway Activation Signatures (iPAS)
}
\description{
Load integrative Pathway Activity Signatures (iPAS)
}
