% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_iPAS_ensemble.R
\name{load_iPAS_ensemble}
\alias{load_iPAS_ensemble}
\title{Load Ensemble Pathway Activity Signatures (iPAS)}
\usage{
load_iPAS_ensemble(genes = c("entrez", "symbol"))
}
\arguments{
\item{genes}{format of rownames for genes (entrez ids or gene symbols)}
}
\value{
a list object with PAS for all cell lines
}
\description{
Load Ensemble Pathway Activity Signatures (iPAS)
}
