% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rnaseq_input_rand}
\alias{rnaseq_input_rand}
\title{Random RNA-seq data to use for examples}
\format{
a matrix containing three input signatures
\describe{
}
}
\usage{
rnaseq_input_rand
}
\description{
This data contained three example input signatures in the form of a matrix
(~26K genes x 3 input signatures). The data was generated by randomly
permuting log fold-change values from three perturbation experiments.
}
\keyword{datasets}
